Add SignedOrder, Order, and ECSignature types to the types package
