+++ packages/connect/package.json
@@ -50,6 +50,7 @@
     "dependencies": {
         "@0xproject/assert": "^0.2.0",
         "@0xproject/json-schemas": "^0.7.14",
+        "@0xproject/types": "^0.3.1",
         "@0xproject/utils": "^0.4.1",
         "isomorphic-fetch": "^2.2.1",
         "lodash": "^4.17.4",
@@ -82,6 +83,6 @@
         "web3-typescript-typings": "^0.10.0"
     },
     "publishConfig": {
-      "access": "public"
+        "access": "public"
     }
 }
